更新项目

+++ examples/modules/math/math.cpp
@@ -23,7 +23,8 @@ bool is_prime(int n) {
     if (n % 2 == 0 || n % 3 == 0) return false;
     
     // 检查 6k ± 1 形式的因子
-    for (int i = 5; i * i <= n; i += 6) {
+    // 使用 long long 转换避免 i * i 在 n 接近 INT_MAX 时溢出
+    for (int i = 5; (long long)i * i <= n; i += 6) {
         if (n % i == 0 || n % (i + 2) == 0) {
             return false;
         }
